requirements.txt
paramiko==2.1.2
pycrypto==2.6.1
ecdsa==0.13
kivy==1.9.1

sftpserver==0.2
pyyaml==3.12
